backend/app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyJWT==2.8.0
bcrypt==4.0.1
flask-restx==1.1.0
APScheduler==3.10.1
pytest==7.4.2
factory-boy==3.3.0
alembic==1.12.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
flask-cors==4.0.0
reportlab==4.0.4
openpyxl==3.1.2
Pillow==10.0.1